+++ MSGraph/changelog.md
@@ -4,6 +4,8 @@
     - returns information about Microsoft Teams team(s)
     - list by default only teams the connected user is joined to
     - switch "ListAll" available for listing all teams in the tenant/organisation
+- New: Command Get-MgaTeamChannel
+    - returns channel(s) from Microsoft Teams team(s)
 - Upd: Manifest information update, while taking ownership for github project from @FredWeinman - thank you Fred!
 - Fix: Parameter dokumentation on various commands for parameter "ResultSize"